Ticker Tape Configuration Updates.  Validation done in Configuration Class.
